reducir la interface a mac_poll y mac_send
