+++ os/arch/board/native/arch_uart.h
@@ -3,7 +3,7 @@
 #define _ARCH_UART_H_
 
 /* Si uart es la salida estandar es aquí donde se incluye stdio */
-#include <stdio.h>
+#include "arch_sys.h"
 
 void uart_init_arch(void);
