+++ os/arch/board/native/native.md
@@ -1,0 +1,15 @@
+# Plataforma nativa
+
+
+## Network
+Pese a que ya está deprecated se ha utilizado **tun/tap** para recrear la interface de red. En la comprensión de uso fue revelador el artículo [Tun/Tap interface tutorial](https://piratelearner.com/en/bookmarks/tuntap-interface-tutorial/14/) el cual recomendamos.
+
+### Instalando en MacOS
+
+Para instalar tun/tap recomendamos brew
+
+``` bash
+    $ brew tap caskroom/cask 
+    $ brew cask install tuntap
+```
+
