+++ os/sys/net/netconf.h
@@ -16,10 +16,16 @@
 
 #ifdef NET_STATIC
 /* Config static by default*/
-#define IP_ADDRESS 192, 168, 1, 115
-#define IP_GATEWAY 192, 168, 1, 1
-#define IP_SUBNET 192, 168, 1, 0
-#define IP_MASK 255, 255, 255, 0
+//#define IP_ADDRESS 192, 168, 1, 115
+//#define IP_GATEWAY 192, 168, 1, 1
+//#define IP_SUBNET 192, 168, 1, 0
+//#define IP_MASK 255, 255, 255, 0
+
+#define IP_ADDRESS 127, 18, 0, 5
+#define IP_GATEWAY 127, 18, 0, 1
+#define IP_SUBNET 127, 18, 0, 0
+#define IP_MASK 255, 255, 0, 0
+
 #endif /* NET_STATIC */
